requirements.txt
flask~=3.1.2
requests~=2.32.3
python-dotenv~=1.1.0
flask-cors
python-dotenv
msal~=1.34.0
microsoft-agents-activity
microsoft-agents-copilotstudio-client
h11~=0.16.0
cryptography~=45.0.5
pip~=25.3
pathlib~=1.0.1
typing_extensions~=4.13.2
emoji~=2.14.1
configparser~=7.2.0
attrs~=25.3.0
argparse~=1.4.0
distro~=1.9.0
django-request~=1.7.1
filelock~=3.18.0
setuptools~=80.9.0
idna~=3.10
multidict~=6.4.4
propcache~=0.3.1
exceptiongroup~=1.3.0
sniffio~=1.3.1
pytest~=8.4.1
aiohttp~=3.12.8
urllib3~=2.4.0
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.2.1
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
yarl~=1.20.0
aiosignal~=1.3.2
frozenlist~=1.6.2
python-multipart~=0.0.18
aiohappyeyeballs~=2.6.1
certifi~=2025.4.26
fastapi~=0.115.12
pydantic~=2.11.5
pydantic_core~=2.33.2
starlette~=0.41.3
responses~=0.25.7
websockets~=15.0.1
anyio~=4.9.0
ujson~=5.11.0
orjson~=3.10.18
isodate~=0.7.2
uvicorn~=0.34.2
wsproto~=1.2.0
watchfiles~=0.20.0
Deprecated~=1.2.18
rich~=14.2.0
chardet~=5.2.0
simplejson~=3.20.1
pycparser~=2.22
PyYAML~=6.0.2
httpx~=0.28.1
PyJWT~=2.10.1
openai~=2.8.1